refactor: コア領域関係のndarray操作ユーティリティを移動
