+++ crates/voicevox_core/src/core/adjust/pre.rs
@@ -1,5 +1,8 @@
 //! 推論の入力の前処理。
 
+use easy_ext::ext;
+use ndarray::{Array1, Array2};
+
 /// 音が途切れてしまうのを避けるworkaround処理。
 // TODO: 改善したらここのpadding処理を取り除く
 pub(crate) fn pad_decoder_feature<const PADDING_FRAME_LENGTH: usize>(
@@ -33,5 +36,13 @@ pub(crate) fn pad_decoder_feature<const PADDING_FRAME_LENGTH: usize>(
             .slice_mut(ndarray::s![.., 0])
             .assign(&ndarray::arr0(1.0));
         ndarray::concatenate![ndarray::Axis(0), padding, phoneme_slice, padding]
+    }
+}
+
+#[ext(Array1ExtForPreProcess)]
+impl<T> Array1<T> {
+    pub(crate) fn into_one_row(self) -> Array2<T> {
+        let n = self.len();
+        self.into_shape_with_order([1, n]).expect("should be ok")
     }
 }
